+++ api/helpers/actions.js
@@ -1,26 +1,29 @@
 "use strict";
-var _ = require('lodash');
-var defaultLog = require('winston').loggers.get('default');
+var { find, isEqual, remove } = require('lodash');
 
 exports.publish = async function (o) {
     return new Promise(function (resolve, reject) {
-        // Object wasn't already published?
-        if (!o.read.includes('public')) {
-            var newReadArray = o.read;
-            newReadArray.push('public');
-            o.read = newReadArray;
-
-            // Remove publish, save then return.
-            resolve(o.save());
-        } else {
-            resolve(o);
+        try {
+            // Object wasn't already published?
+            if (!o.read.includes('public')) {
+                var newReadArray = o.read;
+                newReadArray.push('public');
+                o.read = newReadArray;
+                
+                // Remove publish, save then return.
+                resolve(o.save());
+            } else {
+                resolve(o);
+            }
+        } catch (error) {
+            reject(error);
         }
     });
 };
 
 exports.isPublished = async function (o) {
-    return _.find(o.tags, function (item) {
-        return _.isEqual(item, ['public']);
+    return find(o.tags, function (item) {
+        return isEqual(item, ['public']);
     });
 };
 
@@ -40,8 +43,8 @@ exports.unPublish = async function (o) {
 
 exports.delete = function (o) {
     return new Promise(function (resolve, reject) {
-        _.remove(o.tags, function (item) {
-            return _.isEqual(item, ['public']);
+        remove(o.tags, function (item) {
+            return isEqual(item, ['public']);
         });
         o.isDeleted = true;
         o.markModified('tags');