+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 1.2.0 January 26, 2022
+* Tree shaking lodash dependencies
+* Remove unnecessary console.log calls
+* Graceful error handling for some methods
+* Add details and engagementLabel fields to projects.
+* Add better comments
+* Update some methods with ES6 syntax.
+* Remove unused function params, libraries, variables, etc.
+* Remove old, commented-out code.
+
 ### 1.0.1: November 16, 2021
 * Updating version in package.json
 * security update Force validator 13.7.0 via resolution (LUP-251)